+++ code/BlenderModifier.cpp
@@ -200,9 +200,9 @@ void  BlenderModifier_Mirror :: DoIt(aiNode& out, ConversionData& conv_data,  co
         aiMesh* mesh;
         SceneCombiner::Copy(&mesh,conv_data.meshes[out.mMeshes[i]]);
 
-        const float xs = mir.flag & MirrorModifierData::Flags_AXIS_X ? -1.f : 1.f;
-        const float ys = mir.flag & MirrorModifierData::Flags_AXIS_Y ? -1.f : 1.f;
-        const float zs = mir.flag & MirrorModifierData::Flags_AXIS_Z ? -1.f : 1.f;
+        const double xs = mir.flag & MirrorModifierData::Flags_AXIS_X ? -1.f : 1.f;
+        const double ys = mir.flag & MirrorModifierData::Flags_AXIS_Y ? -1.f : 1.f;
+        const double zs = mir.flag & MirrorModifierData::Flags_AXIS_Z ? -1.f : 1.f;
 
         if (mir.mirror_ob) {
             const aiVector3D center( mir.mirror_ob->obmat[3][0],mir.mirror_ob->obmat[3][1],mir.mirror_ob->obmat[3][2] );
@@ -242,8 +242,8 @@ void  BlenderModifier_Mirror :: DoIt(aiNode& out, ConversionData& conv_data,  co
             }
         }
 
-        const float us = mir.flag & MirrorModifierData::Flags_MIRROR_U ? -1.f : 1.f;
-        const float vs = mir.flag & MirrorModifierData::Flags_MIRROR_V ? -1.f : 1.f;
+        const double us = mir.flag & MirrorModifierData::Flags_MIRROR_U ? -1.f : 1.f;
+        const double vs = mir.flag & MirrorModifierData::Flags_MIRROR_V ? -1.f : 1.f;
 
         for (unsigned int n = 0; mesh->HasTextureCoords(n); ++n) {
             for (unsigned int i = 0; i < mesh->mNumVertices; ++i) {
